fix: implement gamma context and fix issues
