+++ src/Router.tsx
@@ -18,6 +18,13 @@ import {
 import { APP_DEFAULT_ROUTE } from './constants'
 import Maintenance from './pages/Maintenance'
 import { JupWidget } from '@/components/JupWidget'
+import { TraderProvider } from './context/trader_risk_group'
+import { StatsProvider } from './context/stats'
+import { Alignment, Fit, Layout, useRive } from '@rive-app/react-canvas'
+import { Toaster } from 'gfx-component-lib'
+import { RewardsProvider } from '@/context/rewardsContext'
+import { MarketProductGroupProvider } from './context/market_product_group'
+import { GammaProvider } from '@/context/gamma'
 
 const Bridge = lazy(() => import('./pages/Bridge'))
 const GenericNotFound = lazy(() => import('./pages/InvalidUrl'))
@@ -28,12 +35,6 @@ const SSLAnalyticsDashboard = lazy(() => import('./pages/Analytics/ssl/SSLAnalyt
 const LeaderBoard = lazy(() => import('./pages/Stats/LeaderBoard'))
 const Farm = lazy(() => import('./pages/FarmV3/Farm'))
 const FarmV4 = lazy(() => import('./pages/FarmV4/Farm'))
-import { TraderProvider } from './context/trader_risk_group'
-import { StatsProvider } from './context/stats'
-import { Alignment, Fit, Layout, useRive } from '@rive-app/react-canvas'
-import { Toaster } from 'gfx-component-lib'
-import { RewardsProvider } from '@/context/rewardsContext'
-import { MarketProductGroupProvider } from './context/market_product_group'
 
 const CoinGeckoPairs = lazy(() => import('./pages/Analytics/ssl/SSLPairs'))
 const Account = lazy(() => import('./pages/Account/Account'))
